review(#585): three Copilot findings, all valid
**The second undefined cast — the one the first pass missed.** The adapter
truncated each connection's `m_current_speed_*` with
`static_cast<uint64_t>` before it crossed the FFI, which is undefined for a
NaN, an infinity, a negative or an out-of-range value — the same hazard this
PR had just hardened one field away in `round_half_up`, left standing in C++.

Copilot suggested clamping before the cast. Clamping in C++ would close the
UB and get the layering wrong: the truncation is a **derivation**, and this
slice's whole rule is that the adapter transcribes and Rust computes. So the
raw `double`s cross and `kib_per_second` clamps them beside `round_half_up`,
where a test can reach it. That also restores the C++'s actual arithmetic:
the division happened in floating point, so the answer is
`trunc(x / 1024)`, not `trunc(trunc(x) / 1024)` — which the integer POD had
quietly changed it to.

The boundary constant is written out rather than as `u64::MAX as f64`, since
that cast is itself lossy: the nearest `f64` to `u64::MAX` is 2^64, one past
the range, so `>=` against the literal is the correct test.

**A tor peer's address was printed with two ports.** `host` means three
different things across the address arms, and on the tor/i2p arm it is the
whole `network_address::str()` — which already ends in a port — while `port`
is 0. Appending gave `abcdefghijklmnop.onion:18080:0`. The C++ printed that
too, so this is inherited rather than a regression, and it is corrected here
rather than carried (rule 16). The port is appended only when there is one.

**One hex encoder per file, not two.** `hex16` was a hand-rolled per-byte
`format!` in a file that already calls `hex::encode` four times, on a crate
that already depends on it. Deleted. The performance argument is real but
secondary; the reason is that two encoders in one file is a duplicate waiting
to disagree.

Both new behaviours observed red before green: removing the finiteness guard
fails the speed test and only it, and re-appending the port fails the address
test and only it.

`ConnectionFacts` drops its `Eq` derive, which the `f64` fields make
unavailable and which nothing used.

Layout unchanged — `uint64_t` and `double` are both 8 bytes at the same
offset, and the pins on both sides still pass.

Gates: fmt --all, clippy --workspace --all-targets, `shekyl-daemon-rpc` 122
passed, C++ roundtrip + facts shims 39 passed, eight doc gates.

+++ rust/shekyl-daemon-rpc/src/core.rs
@@ -84,7 +84,8 @@ pub enum TxSlot {
 /// derived from these against the one `now` that came back beside the list,
 /// so a connection's reported lifetime and the divisor behind its own
 /// averages cannot come from different seconds.
-#[derive(Debug, Clone, PartialEq, Eq)]
+// No `Eq`: the raw speeds are `f64`, and they are raw on purpose.
+#[derive(Debug, Clone, PartialEq)]
 pub struct ConnectionFacts {
     /// `network_address::str()`: "host:port" for ipv4, the onion for tor.
     pub address: String,
@@ -99,9 +100,11 @@ pub struct ConnectionFacts {
     /// Bytes, absolute.
     pub recv_count: u64,
     pub send_count: u64,
-    /// Bytes/s.
-    pub current_speed_down: u64,
-    pub current_speed_up: u64,
+    /// Bytes/s, raw. Not yet an integer: `static_cast<uint64_t>` of a rate
+    /// estimator's output is undefined for a NaN, an infinity, a negative or
+    /// an out-of-range value, so the truncation happens here, clamped.
+    pub current_speed_down: f64,
+    pub current_speed_up: f64,
     /// The peer's claimed blockchain height.
     pub height: u64,
     pub support_flags: u32,
@@ -118,7 +121,7 @@ pub struct ConnectionFacts {
 }
 
 /// The live connections plus the single instant they were read at.
-#[derive(Debug, Clone, PartialEq, Eq)]
+#[derive(Debug, Clone, PartialEq)]
 pub struct ConnectionsSnapshot {
     /// Unix seconds; one clock read for the whole list.
     pub now: u64,
